+++ source/extensions/omni.isaac.benchmarks/omni/isaac/benchmarks/tests/test_benchmark_lidar.py
@@ -13,23 +13,20 @@ import carb
 from pxr import Gf, UsdGeom
 
 from omni.isaac.core.utils.nucleus import get_assets_root_path
-from omni.isaac.core.utils.stage import add_reference_to_stage, open_stage_async
-from omni.isaac.core.utils.physics import simulate_async
+from omni.isaac.core.utils.stage import open_stage_async
 from omni.isaac.core.utils.rotations import euler_angles_to_quat
 
-# from omni.isaac.core.prims._impl.single_prim_wrapper import set_default_state
-from omni.kit.viewport.utility import get_active_viewport_and_window
-
+from omni.kit.viewport.utility import get_active_viewport, create_viewport_window
 
 import numpy as np
-from ..utils.logger import log_header, log_stamp, get_cpu_info, get_gpu_info
+from ..utils.logger import log_header, get_memory_stats
+from ..utils.helper import delete_all_viewports
 import yaml
 import asyncio
 
 
 class TestBenchmarkLidar(omni.kit.test.AsyncTestCase):
     async def setUp(self):
-        self._timeline = omni.timeline.get_timeline_interface()
         await omni.usd.get_context().new_stage_async()
         await omni.kit.app.get_app().next_update_async()
         pass
@@ -39,8 +36,9 @@ class TestBenchmarkLidar(omni.kit.test.AsyncTestCase):
         pass
 
     # ----------------------------------------------------------------------
-    async def test_benchmark_lidar_sequence(self):
-        test_description = "test up to N lidars for the same scene, no robots"
+    async def test_benchmark_physx_lidar_sequence(self):
+        test_description = "test up to N PhysX lidars for the same scene, no robots"
+        print(test_description)
         stage = omni.usd.get_context().get_stage()
         n_sensor = 3
         n_avg = 5  # number of times to take sample and average
@@ -64,11 +62,7 @@ class TestBenchmarkLidar(omni.kit.test.AsyncTestCase):
             "scene": scene_path,
             "robot": "None",
             "sensors": "Physics Lidar",
-            "fps_raw format": {
-                "row": "data from each viewport n",
-                "column": "samples",
-                "z": "total number of viewports opened",
-            },
+            "fps_raw format": {"row": "data when j sensors are loaded", "column": "samples"},
         }
 
         with open(data_file_path, "a") as f:
@@ -76,15 +70,11 @@ class TestBenchmarkLidar(omni.kit.test.AsyncTestCase):
         f.close()
 
         """
-            data collection loop: for each loop, add a camera and a corresponding viewport
-            to get fps data: averaging across viewports, and across n_avg samples.
+            data collection loop: for each loop, add a lidar
             data formats:
-                @ gpu memory: [nxm],  n = # of gpus, m = # of total cameras added
-                @ cpu memory: [mx1], m = # of total cameras added
-                @ fps_raw: [n_sensors x n_avg], 
-                @ fps_mean: [mx1], m = # of total cameras added, (e.g. the i'th element = the fps average across n_avg samples of having i viewports/cameras open) 
-                @ fps_var_samples: [mx1], variance across samples (if large means performance isn't stable)
-                @ fps_var_viewports: [mx1], variance across viewports (if large then viewports are rendering at different rates)
+                @ gpu memory: [nxm],  n = # of gpus, m = # of total sensors added
+                @ cpu memory: [mx1], m = # of total sensors added
+                @ fps_raw: [n_sensors x n_avg], row = j'th row means j sensors are loaded when taken data in this row, column = samples
         """
 
         def add_lidar(prim_path, translation=Gf.Vec3f(0, 0, 0), orientation=Gf.Vec4f(0, 0, 0, 0)):
@@ -114,13 +104,19 @@ class TestBenchmarkLidar(omni.kit.test.AsyncTestCase):
             lidar_prim.GetAttribute("xformOp:translate").Set(translation)
             lidar_prim.GetAttribute("xformOp:orient").Set(orientation)
 
+        timeline = omni.timeline.get_timeline_interface()
+        timeline.play()
         # data arrays
         fps_raw = np.zeros([n_sensor, n_avg])
-        fps_mean = np.zeros([n_sensor, 1])
-        fps_var_samples = np.zeros([n_sensor, 1])
-        fps_var_viewports = np.zeros([n_sensor, 1])
-
+        cpu_raw = np.zeros([n_sensor, n_avg])
+        gpu_raw = np.zeros([n_sensor, n_avg])
         for i in range(n_sensor):
+
+            ## Delete current viewports and open a new one for a new resolution
+            delete_all_viewports()
+            await omni.kit.app.get_app().next_update_async()
+            create_viewport_window(name="Viewport")
+            await omni.kit.app.get_app().next_update_async()
 
             # add a sensor on stage
             sensor_path = "/World/Lidar_" + str(i)
@@ -129,31 +125,30 @@ class TestBenchmarkLidar(omni.kit.test.AsyncTestCase):
             sensor_orientation = Gf.Quatf(q[0], q[1], q[2], q[3])
             add_lidar(prim_path=sensor_path, translation=sensor_translation, orientation=sensor_orientation)
             # Run for a second
-            self._timeline.play()
-            await simulate_async(1.0)
-            self._timeline.pause()
+            await asyncio.sleep(1.0)
             print("lidar {} added at {}".format(i, sensor_path))
 
-            # take a sample one second apart (#TODO: or should it be take a sample that last n seconds and then average)
+            # take a sample one second apart
             for j in range(n_avg):
                 for s in range(60):
                     await omni.kit.app.get_app().next_update_async()
 
-                viewport_api, active_window = get_active_viewport_and_window()
-                fps_raw[i, j] = viewport_api.fps
-                # print("fps",viewport_window.fps)
+                # get performance data
+                viewport_window = get_active_viewport()
+                fps_raw[i, j] = viewport_window.fps
+                # get memory data
+                memory_usage = get_memory_stats()
+                cpu_raw[i, j] = memory_usage["System Memory"]["RAM"]
+                gpu_raw[i, j] = memory_usage["System Memory"]["VRAM"]
 
             # end of taking multiple samples
-            # get memory data
-            cpu_info = get_cpu_info()
-            gpu_info = get_gpu_info()
 
         # end of adding cameras
-        per_res_data = {"data": {"fps_raw": str(fps_raw), "cpu_info": cpu_info, "gpu_info": gpu_info}}
+        per_res_data = {"data": {"fps_raw": str(fps_raw), "cpu_info": str(cpu_raw), "gpu_info": str(gpu_raw)}}
         with open(data_file_path, "a") as f:
             yaml.safe_dump(per_res_data, f)
         f.close()
 
         print("fps_raw: ", fps_raw)
-        print("cpu_info: ", cpu_info)
-        print("gpu_info: ", gpu_info)
+        print("cpu_info: ", cpu_raw)
+        print("gpu_info: ", gpu_raw)